refactoring: style 코드 컴포넌트와 분리
