+++ src/components/commons/Logo.tsx
@@ -1,19 +1,6 @@
 import React from "react";
-import styled from "styled-components";
 import logo from "assets/images/logo.png";
-
-const StyledLogo = styled.div.attrs({ className: "logo" })`
-  display: inline-block;
-  img {
-    height: 3rem;
-    width: 10rem;
-    display: inline-block;
-    ${({ theme }) => theme.media.mobile`        
-      height: 2rem;
-      width: 6.7rem;
-    `}
-  }
-`;
+import { StyledLogo } from "styles/components/commons/Logo.style";
 
 const Logo = (): JSX.Element => {
   return (
