+++ src/styles/components/commons/Layout.style.ts
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const StyledLayout = styled.div`
+  width: 100%;
+  height: 100%;
+`;
